requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.55.3
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv
httpx==0.27.2